delete article test

+++ src/main/java/de/ait/platform/article/service/ArticleServiceImp.java
@@ -68,6 +68,7 @@ public class ArticleServiceImp implements ArticleService {
         Predicate<Article> predicateByTitle =
                 (title.equals("")) ? a-> true:  article -> article.getTitle().equalsIgnoreCase(title);
         List<Article> articleList = repository.findAll().stream().filter(predicateByTitle).toList();
+        System.out.println(articleList);
         if (articleList.isEmpty()) {
             throw new ArticleNotFound("Article with title: " + title + " not found");
         }
@@ -171,19 +172,20 @@ public class ArticleServiceImp implements ArticleService {
     @Override
     public ResponseArticle deleteArticle(Long id) {
         Optional<Article> foundedArticle = repository.findById(id);
-        if (foundedArticle.isPresent()) {
-            Set<Comment> comments = foundedArticle.get().getComments();
-            for (Comment comment : comments) {
-                comment.setArticle(null);
-                comment.setUser(null);
+        if (foundedArticle.isPresent()){
+            if (foundedArticle.get().getComments() != null && !foundedArticle.get().getComments().isEmpty()){
+                for (Comment comment : foundedArticle.get().getComments()) {
+                    comment.setArticle(null);
+                    comment.setUser(null);
+                }
+                foundedArticle.get().setComments(new HashSet<>());
+                repository.deleteById(id);
             }
-            foundedArticle.get().setComments(new HashSet<>());
-            repository.deleteById(id);
-            return mapper.map(foundedArticle, ResponseArticle.class);
         }
         else {
             throw new ArticleNotFound("Article with id: " + id + "not found");
         }
 
+        return mapper.map(foundedArticle, ResponseArticle.class);
     }
 }
